feat: Integrate `label.data` into label component
Integrate `label.data` into the label component and populate it with
demo label data. This modification enhances maintainability by centralizing
label-related data within the component.
